refactor(core): split client remote ssh_transfer into submodules
